add delete feature

+++ scripts/index.js
@@ -5,7 +5,11 @@ const argv = yargs(hideBin(process.argv)).argv;
 
 // 1. Check if user is at root folder
 const { isAtRootFolder } = require('./utils');
-const { getAndValidateUserInput, isIncludeTestFiles } = require('./cli');
+const {
+  getAndValidateUserInput,
+  isIncludeTestFiles,
+  getDeleteFilesOptions,
+} = require('./cli');
 const CodeFilesHandler = require('./CodeFilesHandler');
 const TestFilesHandler = require('./TestFilesHandler');
 if (!isAtRootFolder()) return;
@@ -21,7 +25,16 @@ const testFilesHandler = new TestFilesHandler(singular, plural);
 // 4. handle delete files case
 
 if (argv.delete || argv.d || argv._.includes('delete')) {
-  codeFilesHandler.deleteCodeFiles();
+  const { deleteCodeFiles, deleteTestFiles } = getDeleteFilesOptions();
+
+  if (deleteCodeFiles) {
+    codeFilesHandler.deleteCodeFiles();
+  }
+
+  if (deleteTestFiles) {
+    testFilesHandler.deleteTestFiles();
+  }
+
   return;
 }
